r389: Added a man-page (Andy Piper). Changed `install.sh' from bash to sh (Bernard Jungen).

+++ ROX-Filer/src/remote.c
@@ -70,7 +70,7 @@ gboolean remote_init(GString *to_open, gboolean new_copy)
 	GtkWidget		*ipc_window;
 
 	unique_id = g_strdup_printf("_ROX_FILER_%d_%s_%s",
-				euid, VERSION, our_host_name());
+				(int) euid, VERSION, our_host_name());
 	filer_atom = gdk_atom_intern(unique_id, FALSE);
 	g_free(unique_id);
